refactor: Dao의 기능을 인터페이스로 분리
