+++ src/main/java/wooteco/subway/dao/StationDao.java
@@ -1,62 +1,18 @@
 package wooteco.subway.dao;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
 import java.util.List;
-
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
 
 import wooteco.subway.domain.Station;
 
-@Repository
-public class StationDao {
-    private final JdbcTemplate jdbcTemplate;
+public interface StationDao {
+    Station save(Station station);
 
-    public StationDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
+    boolean existByName(String name);
 
-    public Station save(Station station) {
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        String sql = "INSERT INTO STATION(name) VALUES(?)";
-        String name = station.getName();
-        jdbcTemplate.update((Connection conn) -> {
-            PreparedStatement pstmt = conn.prepareStatement(sql, new String[] {"id"});
-            pstmt.setString(1, name);
-            return pstmt;
-        }, keyHolder);
+    List<Station> findAll();
 
-        Long id = keyHolder.getKey().longValue();
+    void delete(Long id);
 
-        return new Station(id, name);
-    }
+    boolean existById(Long id);
 
-    public boolean existByName(String name) {
-        String sql = "SELECT EXISTS(SELECT id FROM STATION WHERE name = ?) AS SUCCESS";
-        return jdbcTemplate.queryForObject(sql, Boolean.class, name);
-    }
-
-    public List<Station> findAll() {
-        String sql = "SELECT id, name FROM STATION";
-        return jdbcTemplate.query(sql, (resultSet, rowNum) -> {
-            Station station = new Station(
-                resultSet.getLong("id"),
-                resultSet.getString("name")
-            );
-            return station;
-        });
-    }
-
-    public void delete(Long id) {
-        String sql = "DELETE FROM STATION WHERE id = ?";
-        jdbcTemplate.update(sql, id);
-    }
-
-    public boolean existById(Long id) {
-        String sql = "SELECT EXISTS(SELECT * FROM STATION WHERE id = ?) AS SUCCESS";
-        return jdbcTemplate.queryForObject(sql, Boolean.class, id);
-    }
 }
